refactor(view): share list prefix validation (#21)
- Remove duplicated raw-prefix validation from client implementations
- Keep C++20 list semantics and Status behavior unchanged
- Reduce new-code duplication without changing public APIs
